--- src/main/java/nextstep/member/domain/GithubOAuth2Client.java
@@ -1,12 +1,0 @@
-package nextstep.member.domain;
-
-import nextstep.member.payload.AccessTokenResponse;
-import nextstep.member.payload.GithubUserInfoResponse;
-
-public interface GithubOAuth2Client {
-
-    AccessTokenResponse getAccessToken(String code);
-
-    GithubUserInfoResponse getUserInfo(String accessToken);
-
-}
