+++ src/main/java/nextstep/security/oauth2/GithubOAuthProperty.java
@@ -1,4 +1,4 @@
-package nextstep.member.domain;
+package nextstep.security.oauth2;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
