+++ src/main/java/nextstep/security/oauth2/GithubOauth2Client.java
@@ -1,0 +1,6 @@
+package nextstep.security.oauth2;
+
+public interface GithubOauth2Client extends OAuth2Client {
+    GithubUserInfoResponse getUserInfo(String accessToken);
+
+}
