+++ src/main/java/nextstep/security/oauth2/GithubUserInfoResponse.java
@@ -1,4 +1,4 @@
-package nextstep.member.payload;
+package nextstep.security.oauth2;
 
 public class GithubUserInfoResponse {
