+++ src/main/java/nextstep/security/oauth2/OAuth2Client.java
@@ -1,0 +1,9 @@
+package nextstep.security.oauth2;
+
+public interface OAuth2Client {
+
+    AccessTokenResponse getAccessToken(String code);
+
+    Object getUserInfo(String accessToken);
+
+}
